Split session handling into separate class
